Fix tests

+++ api/internal/recipe/dto.go
@@ -7,6 +7,6 @@ type CreateRecipeRequest struct {
 
 type CreateIngredientUsageRequest struct {
 	IngredientID string  `json:"ingredient_id" validate:"required,uuid4"`
-	Quantity     float32 `json:"quantity" validate:"required,gt=0"`
+	Quantity     float64 `json:"quantity" validate:"required,gt=0"`
 	Unit         int     `json:"unit" validate:"required"`
 }
